Initial work on implementing multiple AnimatedOrphans support
