+++ addon/templates/components/-animated-orphans.hbs
@@ -1,0 +1,1 @@
+{{yield}}
