+++ addon/templates/components/animated-orphans.hbs
@@ -1,1 +1,2 @@
+{{-animated-orphans parent=this}}
 {{yield}}
