+++ app/components/-animated-orphans.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-animated/components/-animated-orphans';
